requirements.txt
fastapi
uvicorn
python-dotenv
PyGithub
polars
google.generativeai
gitingest
Colorama
Click